fix issue #81 : refactor code

+++ migrations/2020-04-18-085031_create_posts/up.sql
@@ -10,8 +10,8 @@ CREATE TABLE posts (
     deleted_at TIMESTAMP NULL,
     hidden_at TIMESTAMP NULL,
     locked_at TIMESTAMP NULL,
-    votes INT UNSIGNED NOT NULL DEFAULT 0,
-    score INT NOT NULL DEFAULT 0,
+    votes BIGINT UNSIGNED NOT NULL DEFAULT 0,
+    score BIGINT NOT NULL DEFAULT 0,
     PRIMARY KEY (id),
     CONSTRAINT fk_posts_authorid FOREIGN KEY (author_id) REFERENCES users(id),
     CONSTRAINT posttype_values CHECK (post_type in ("Poll", "Idea", "Info", "Decisional", "Discussion"))
